requirements.txt
fastapi
uvicorn
sqlalchemy
scikit-learn
joblib
python-dotenv
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-text-splitters
faiss-cpu
sentence-transformers
transformers
torch
